refactor(#zimic): move shared files to dedicated folder to avoid import pollution
