+++ packages/zimic/src/cli/browser/shared/constants.ts
@@ -1,0 +1,1 @@
+export const SERVICE_WORKER_FILE_NAME = 'mockServiceWorker.js';
