+++ packages/zimic/src/interceptor/http/interceptor/__tests__/shared/methods/get.ts
