+++ packages/zimic/src/interceptor/http/interceptor/__tests__/shared/methods/head.ts
