+++ packages/zimic/src/interceptor/http/interceptorWorker/__tests__/BrowserHttpinterceptorWorker.browser.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, expectTypeOf, it } from 'vitest';
 import BrowserHttpInterceptorWorker from '../BrowserHttpInterceptorWorker';
 import HttpInterceptorWorker from '../HttpInterceptorWorker';
 import { BrowserHttpWorker } from '../types';
-import { declareSharedHttpInterceptorWorkerTests } from './sharedTests';
+import { declareSharedHttpInterceptorWorkerTests } from './shared/workerTests';
 
 describe('BrowserHttpInterceptorWorker', () => {
   it('should initialize using the Browser.js MSW server', () => {
