+++ packages/zimic/src/interceptor/http/interceptorWorker/__tests__/NodeHttpinterceptorWorker.node.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, expectTypeOf, it } from 'vitest';
 import HttpInterceptorWorker from '../HttpInterceptorWorker';
 import NodeHttpInterceptorWorker from '../NodeHttpInterceptorWorker';
 import { NodeHttpWorker } from '../types';
-import { declareSharedHttpInterceptorWorkerTests } from './sharedTests';
+import { declareSharedHttpInterceptorWorkerTests } from './shared/workerTests';
 
 describe('NodeHttpInterceptorWorker', () => {
   it('should initialize using the Node.js MSW server', () => {
